+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 更改日志 | CHANGE LOG
 
 - 自动加载块属性 `custom-data-assets` 中设置的资源文件 | The resource file set in the block attribute `custom-data-assets` is automatically loaded.
+- 添加保存到思源按钮 | Add the Save to Source button.